06.08.2024 Добавлены свойства товара и характеристик в отдельную таблицу

+++ src/main/java/main/model/propertyes/PropertyReposytory.java
@@ -2,7 +2,7 @@ package main.model.propertyes;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface PropertyReposytory extends CrudRepository<Property, Integer> {
+public interface PropertyReposytory extends CrudRepository<Property, Long> {
 
     Property findById1c(String id1c);
